Ticket_Management_system
Because this Project is using RMI and hibernate you have to first to connect your database with hibernate on this project then go on the part of the server and Run the server before you run the project on the client side after running the server and the client is on running the admin username and password are embedded in source code of the loginApp that's where the admin will be managing the user's who will be registering the ticket management systems. On the client side to start the project you have to first go on the view folder then run the process.java to start the project.